# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HeartEcho API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserSubscriptionPlan = typeof UserSubscriptionPlan[keyof typeof UserSubscriptionPlan];


export const UserSubscriptionPlan = {
  saving_phase: 'saving_phase',
  legacy_connected: 'legacy_connected',
} as const;

export type UserLanguage = typeof UserLanguage[keyof typeof UserLanguage];


export const UserLanguage = {
  en: 'en',
  es: 'es',
  hi: 'hi',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  subscriptionPlan: UserSubscriptionPlan;
  subscriptionPrice?: string;
  language: UserLanguage;
  createdAt: string;
}

export type UserRegistrationLanguage = typeof UserRegistrationLanguage[keyof typeof UserRegistrationLanguage];


export const UserRegistrationLanguage = {
  en: 'en',
  es: 'es',
  hi: 'hi',
} as const;

export interface UserRegistration {
  name: string;
  email: string;
  password: string;
  language: UserRegistrationLanguage;
}

export type SubscriptionUpdateSubscriptionPlan = typeof SubscriptionUpdateSubscriptionPlan[keyof typeof SubscriptionUpdateSubscriptionPlan];


export const SubscriptionUpdateSubscriptionPlan = {
  saving_phase: 'saving_phase',
  legacy_connected: 'legacy_connected',
} as const;

export interface SubscriptionUpdate {
  subscriptionPlan: SubscriptionUpdateSubscriptionPlan;
}

export interface Nominee {
  id: number;
  name: string;
  email: string;
  phone: string;
  userId: number;
}

export interface NomineeInput {
  name: string;
  email: string;
  phone: string;
}

export type MemoryProfileProfileType = typeof MemoryProfileProfileType[keyof typeof MemoryProfileProfileType];


export const MemoryProfileProfileType = {
  living: 'living',
  ancestor: 'ancestor',
} as const;

export interface MemoryProfile {
  id: number;
  seniorName: string;
  hometown: string;
  favoriteThings: string;
  lifeMantra: string;
  profileType: MemoryProfileProfileType;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  voiceSnippetUrl?: string | null;
  userId: number;
  storyCount: number;
  minutesUsedThisMonth: number;
  createdAt: string;
}

export type MemoryProfileInputProfileType = typeof MemoryProfileInputProfileType[keyof typeof MemoryProfileInputProfileType];


export const MemoryProfileInputProfileType = {
  living: 'living',
  ancestor: 'ancestor',
} as const;

export interface MemoryProfileInput {
  seniorName: string;
  hometown: string;
  favoriteThings: string;
  lifeMantra: string;
  profileType: MemoryProfileInputProfileType;
}

export interface MemoryProfileUpdate {
  seniorName?: string;
  hometown?: string;
  favoriteThings?: string;
  lifeMantra?: string;
  photoUrl?: string;
  voiceSnippetUrl?: string;
}

export interface Recording {
  id: number;
  title: string;
  durationSeconds: number;
  /** @nullable */
  audioUrl?: string | null;
  chapter: string;
  /** @nullable */
  promptId?: number | null;
  profileId: number;
  createdAt: string;
}

export interface RecordingInput {
  title: string;
  durationSeconds: number;
  audioUrl?: string;
  chapter: string;
  promptId?: number;
  profileId: number;
}

export interface RecordingUpdate {
  title?: string;
  audioUrl?: string;
  durationSeconds?: number;
}

export type TimeCapsuleOccasion = typeof TimeCapsuleOccasion[keyof typeof TimeCapsuleOccasion];


export const TimeCapsuleOccasion = {
  birthday: 'birthday',
  anniversary: 'anniversary',
  holiday: 'holiday',
  custom: 'custom',
} as const;

export interface TimeCapsule {
  id: number;
  title: string;
  occasion: TimeCapsuleOccasion;
  scheduledDate: string;
  /** @nullable */
  audioUrl?: string | null;
  isLocked: boolean;
  profileId: number;
  createdAt: string;
}

export type TimeCapsuleInputOccasion = typeof TimeCapsuleInputOccasion[keyof typeof TimeCapsuleInputOccasion];


export const TimeCapsuleInputOccasion = {
  birthday: 'birthday',
  anniversary: 'anniversary',
  holiday: 'holiday',
  custom: 'custom',
} as const;

export interface TimeCapsuleInput {
  title: string;
  occasion: TimeCapsuleInputOccasion;
  scheduledDate: string;
  audioUrl?: string;
  profileId: number;
}

export interface StoryPrompt {
  id: number;
  question: string;
  category: string;
  weekNumber: number;
  isActive: boolean;
}

export type ArchiveItemItemType = typeof ArchiveItemItemType[keyof typeof ArchiveItemItemType];


export const ArchiveItemItemType = {
  photo: 'photo',
  voice_snippet: 'voice_snippet',
} as const;

export interface ArchiveItem {
  id: number;
  itemType: ArchiveItemItemType;
  fileUrl: string;
  /** @nullable */
  description?: string | null;
  profileId: number;
  uploadedAt: string;
}

export type ArchiveItemInputItemType = typeof ArchiveItemInputItemType[keyof typeof ArchiveItemInputItemType];


export const ArchiveItemInputItemType = {
  photo: 'photo',
  voice_snippet: 'voice_snippet',
} as const;

export interface ArchiveItemInput {
  itemType: ArchiveItemInputItemType;
  fileUrl: string;
  description?: string;
  profileId: number;
}

export interface Questionnaire {
  id: number;
  profileId: number;
  /** @nullable */
  birthYear?: number | null;
  /** @nullable */
  occupation?: string | null;
  /** @nullable */
  hometown?: string | null;
  /** @nullable */
  personality?: string | null;
  /** @nullable */
  favoriteFoods?: string | null;
  /** @nullable */
  traditions?: string | null;
  /** @nullable */
  lifePhilosophy?: string | null;
  /** @nullable */
  memorableStories?: string | null;
  submittedAt: string;
}

export interface QuestionnaireInput {
  profileId: number;
  birthYear?: number;
  occupation?: string;
  hometown?: string;
  personality?: string;
  favoriteFoods?: string;
  traditions?: string;
  lifePhilosophy?: string;
  memorableStories?: string;
}

export type CallSessionSessionType = typeof CallSessionSessionType[keyof typeof CallSessionSessionType];


export const CallSessionSessionType = {
  video: 'video',
  audio: 'audio',
} as const;

export interface CallSession {
  id: number;
  sessionType: CallSessionSessionType;
  durationMinutes: number;
  profileId: number;
  startedAt: string;
}

export type CallSessionInputSessionType = typeof CallSessionInputSessionType[keyof typeof CallSessionInputSessionType];


export const CallSessionInputSessionType = {
  video: 'video',
  audio: 'audio',
} as const;

export interface CallSessionInput {
  sessionType: CallSessionInputSessionType;
  durationMinutes: number;
  profileId: number;
}

export interface DashboardSummary {
  totalRecordings: number;
  totalStoriesMinutes: number;
  activeProfiles: number;
  timeCapsulesPending: number;
  minutesUsedThisMonth: number;
  minutesIncluded: number;
  subscriptionPlan: string;
  recentRecordings: Recording[];
}

export type MilestoneBadgeTier = typeof MilestoneBadgeTier[keyof typeof MilestoneBadgeTier];


export const MilestoneBadgeTier = {
  gold: 'gold',
  silver: 'silver',
  bronze: 'bronze',
} as const;

export interface Milestone {
  label: string;
  achievedAt: string;
  badgeTier: MilestoneBadgeTier;
}

export interface ProfileBadges {
  profileId: number;
  goldBadges: number;
  silverBadges: number;
  bronzeBadges: number;
  totalStories: number;
  currentStreak: number;
  milestones: Milestone[];
}

export type ListRecordingsParams = {
profileId?: number;
};

export type ListTimeCapsulesParams = {
profileId?: number;
};

export type ListArchiveItemsParams = {
profileId?: number;
};

export type ListSessionsParams = {
profileId?: number;
};