# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HeartEcho API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/subscription:
    patch:
      operationId: updateSubscription
      tags: [users]
      summary: Update user subscription plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionUpdate"
      responses:
        "200":
          description: Subscription updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        subscriptionPlan:
          type: string
          enum: [saving_phase, legacy_connected]
        subscriptionPrice:
          type: string
        language:
          type: string
          enum: [en, es, hi]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - subscriptionPlan
        - language
        - createdAt
    UserRegistration:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        language:
          type: string
          enum: [en, es, hi]
      required:
        - name
        - email
        - password
        - language
    SubscriptionUpdate:
      type: object
      properties:
        subscriptionPlan:
          type: string
          enum: [saving_phase, legacy_connected]
      required:
        - subscriptionPlan
...