# lib/api-zod/src/generated/types/subscriptionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HeartEcho API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionUpdateSubscriptionPlan } from './subscriptionUpdateSubscriptionPlan';

export interface SubscriptionUpdate {
  subscriptionPlan: SubscriptionUpdateSubscriptionPlan;
}