# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HeartEcho API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserLanguage } from './userLanguage';
import type { UserSubscriptionPlan } from './userSubscriptionPlan';

export interface User {
  id: number;
  name: string;
  email: string;
  subscriptionPlan: UserSubscriptionPlan;
  subscriptionPrice?: string;
  language: UserLanguage;
  createdAt: string;
}